requirements.txt
python-dotenv
kaggle
pandas
numpy
matplotlib
statistics
seaborn
catboost
pyyaml
